refactored sessions in discordJs
