+++ bot/jsonDb/sessionAddUserRequest.json
@@ -216,6 +216,71 @@
             "session_channel_id": "1031474793045819443",
             "requested": [],
             "denied": []
+        },
+        {
+            "session_channel_id": "1035570174973194354",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035576067139846144",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035576970957828116",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035602736403726438",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035603247152504832",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035605796890873987",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035606011756675082",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035606166954319974",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035607021543764048",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035607927869952030",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035608624984248430",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035611567196213249",
+            "requested": [],
+            "denied": []
+        },
+        {
+            "session_channel_id": "1035618042811523093",
+            "requested": [],
+            "denied": []
         }
     ]
 }
